[Behat] Extract Tax transformer
