+++ etc/behat/services/contexts.xml
@@ -34,6 +34,7 @@
         <parameter key="sylius.behat.context.transform.payment.class">Sylius\Behat\Context\Transform\PaymentContext</parameter>
         <parameter key="sylius.behat.context.transform.product.class">Sylius\Behat\Context\Transform\ProductContext</parameter>
         <parameter key="sylius.behat.context.transform.shipping.class">Sylius\Behat\Context\Transform\ShippingContext</parameter>
+        <parameter key="sylius.behat.context.transform.tax.class">Sylius\Behat\Context\Transform\TaxContext</parameter>
 
         <parameter key="sylius.behat.context.hook.doctrine_orm.class">Sylius\Behat\Context\Hook\DoctrineORMContext</parameter>
 
@@ -191,6 +192,11 @@
 
         <service id="sylius.behat.context.transform.shipping" class="%sylius.behat.context.transform.shipping.class%" scope="scenario">
             <argument type="service" id="sylius.repository.shipping_method" container="symfony" />
+            <tag name="sylius.behat.context" />
+        </service>
+
+        <service id="sylius.behat.context.transform.tax" class="%sylius.behat.context.transform.tax.class%" scope="scenario">
+            <argument type="service" id="sylius.repository.tax_category" container="symfony" />
             <tag name="sylius.behat.context" />
         </service>
