+++ etc/behat/suites/ui_cart.yml
@@ -22,6 +22,7 @@ default:
                 - sylius.behat.context.transform.shared_storage
                 - sylius.behat.context.transform.product
                 - sylius.behat.context.transform.shipping
+                - sylius.behat.context.transform.tax
 
                 - sylius.behat.context.ui.checkout
                 - sylius.behat.context.ui.cart
